+++ web/components/index/MainContent.tsx
@@ -6,6 +6,7 @@ import { motion } from 'framer-motion';
 import GenericImage from './GenericImage';
 import StrawberryLinks from './StrawberryLinks';
 import KirakiraEffect from './KirakiraEffect';
+import { MembersLinks } from './MembersLinks';
 
 import { useAnimationWithHover } from '../../hooks/hooks_index'
 
@@ -19,33 +20,8 @@ const MainContent = () => {
   const [isHovered, setIsHovered] = useState(false);
 
   const { controls, animationVariants } = useAnimationWithHover(isHovered);
-
-    // 各段ごとのリンク情報を定義
-    const links1 = [
-      { href: "/members/kato", text: "kato" },
-      { href: "/members/mick", text: "mick" },
-      { href: "/members/nori", text: "nori" },
-      { href: "/members/konisi", text: "konisi" },
-    ];
   
-    const links2 = [
-      { href: "/members/knnts", text: "kana nuts"  },
-      { href: "/members/mchan", text: "Mchan"  },
-    ];
-  
-    const links3 = [
-      { href: "/members/natsu", text: "natsu" },
-      { href: "/members/rei", text: "rei" },
-      { href: "/members/sima9z", text: "sima9z" },
-      { href: "/members/hase", text: "hase" }
-    ];
-  
-    const links4 = [
-      { href: "/members/yu-ki", text: "yu-ki" },
-      { href: "/members/shino", text: "shino" },
-      { href: "/members/rio", text: "rio" },
-      { href: "/members/syuji", text: "syuji" },
-    ];
+  const { links1, links2, links3, links4 } = MembersLinks();
 
   return (
     <>
