+++ web/components/index/MembersLinks.ts
@@ -1,0 +1,31 @@
+export const MembersLinks = () =>{
+  // 各段ごとのリンク情報を定義
+  const links1 = [
+    { href: "/members/kato", text: "kato" },
+    { href: "/members/mick", text: "mick" },
+    { href: "/members/nori", text: "nori" },
+    { href: "/members/konisi", text: "konisi" },
+  ];
+
+  const links2 = [
+    { href: "/members/knnts", text: "kana nuts"  },
+    { href: "/members/mchan", text: "Mchan"  },
+  ];
+
+  const links3 = [
+    { href: "/members/natsu", text: "natsu" },
+    { href: "/members/rei", text: "rei" },
+    { href: "/members/sima9z", text: "sima9z" },
+    { href: "/members/hase", text: "hase" }
+  ];
+
+  const links4 = [
+    { href: "/members/yu-ki", text: "yu-ki" },
+    { href: "/members/shino", text: "shino" },
+    { href: "/members/rio", text: "rio" },
+    { href: "/members/syuji", text: "syuji" },
+  ];
+  
+  return { links1, links2, links3, links4 };
+};
+
